Engine: more documentation of hardcoded API

+++ Engine/scenes/level/lvl_npc.h
@@ -321,7 +321,7 @@ public:
     int  lua_frameDelay();
     void lua_setFrameDelay(int ms);
     int lua_activate_neighbours();
-    LVL_Npc *lua_spawnNPC(int npcID, int sp_type, int sp_dir, bool reSpawnable = false);
+    LVL_Npc *lua_spawnNPC(unsigned long npcID, int sp_type, int sp_dir, bool reSpawnable = false);
 
     inline void lua_setGfxOffsetX(double x)
     {
